refactor: move out runtime paths
